lua/film.c - return nil instead of Lua error
Function film_getnum() is the indexing function for a Lua table of type
dt_lua_film_t (film roll). If the index exceeds the number of images
contained in a film roll, the function should return nil instead of
generating a Lua error to comply to the requirements of Lua tables.